add helpers namespace
